count letters and words
I added two functions that count the number of letters and words

I use the Coleman-Liau index (index = 0.0588 * L - 0.296 * S - 15.8) to calculate the grade level of sentences. L is the average number of letters per 100 words in the text, and S is the average number of sentences per 100 words in the text. Thus, I needed to add the functions to count the number of letters and words. 

I will create a function that counts the number of sentences.