fix count sentences and add grade level
I fixed the count-sentences function and added the evaluate-grade-level function

I forgot to add "I" and "?" as the signs of the end of sentences in the previous commit, so I added them. 

I also added the function to evaluate grade level, using the Coleman-Liau index. 

These fix/addition make this cs50 assignment complete.